My Work As A Designer In This Project.
I've redesigned the entire original layout of this website from when we created it back in 3rd year of high school. I've added gradient effects to the navigation bar, along with an entire new color scheme. I redesigned the layout of buttons and added animations them. I've also changed the typography in the website by redesigning the texts by changing text color and fonts. I have also redesigned the main focus of the project, the chart that is made via javascript when ran, it originally used to be a plain white and blue color scheme but I recolored and resized it for a better look.